Move env packages to the internal package
