+++ compiler/ballerina-lang/src/main/java/io/ballerina/projects/internal/environment/DefaultPackageResolver.java
@@ -15,11 +15,13 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.projects.env;
+
+package io.ballerina.projects.internal.environment;
 
 import io.ballerina.projects.Module;
 import io.ballerina.projects.Package;
 import io.ballerina.projects.PackageId;
+import io.ballerina.projects.Project;
 import io.ballerina.projects.SemanticVersion;
 import io.ballerina.projects.environment.GlobalPackageCache;
 import io.ballerina.projects.environment.ModuleLoadRequest;
@@ -34,25 +36,36 @@ import java.util.List;
 import java.util.Optional;
 
 /**
- * Resolves lang lib packages.
+ * Default Package resolver for Ballerina project.
  *
  * @since 2.0.0
  */
-public class LangLibResolver extends PackageResolver {
+public class DefaultPackageResolver extends PackageResolver {
+    private final Project project;
     private final Repository distCache;
     private final GlobalPackageCache globalPackageCache;
 
-    public LangLibResolver(Repository distCache, GlobalPackageCache globalPackageCache) {
+    public DefaultPackageResolver(Project project, Repository distCache, GlobalPackageCache globalPackageCache) {
+        this.project = project;
         this.distCache = distCache;
         this.globalPackageCache = globalPackageCache;
     }
 
-    @Override
     public Collection<ModuleLoadResponse> loadPackages(Collection<ModuleLoadRequest> moduleLoadRequests) {
         // TODO This is a dummy implementation that checks only the current package in the project.
         List<ModuleLoadResponse> modLoadResponses = new ArrayList<>();
+        Package currentPkg = this.project.currentPackage();
         for (ModuleLoadRequest modLoadRequest : moduleLoadRequests) {
-            Module module = loadFromCache(modLoadRequest);
+            Module module = null;
+            if (currentPkg.packageName().equals(modLoadRequest.packageName())) {
+                module = currentPkg.module(modLoadRequest.moduleName());
+            }
+
+            if (module == null) {
+                // Try to get from the already loaded packages
+                module = loadFromCache(modLoadRequest);
+            }
+
             if (module == null) {
                 module = loadFromDistributionCache(modLoadRequest);
             }
@@ -66,8 +79,10 @@ public class LangLibResolver extends PackageResolver {
         return modLoadResponses;
     }
 
-    @Override
     public Package getPackage(PackageId packageId) {
+        if (project.currentPackage().packageId() == packageId) {
+            return project.currentPackage();
+        }
         return globalPackageCache.get(packageId);
     }
 
@@ -83,12 +98,10 @@ public class LangLibResolver extends PackageResolver {
                 return null;
             }
             SemanticVersion latest = findlatest(packageVersions);
-            loadRequest = new PackageLoadRequest(loadRequest.orgName()
-                    .orElse(null), loadRequest.packageName(), latest);
+            loadRequest = new PackageLoadRequest(loadRequest.orgName().orElse(null), loadRequest.packageName(), latest);
         }
 
         Optional<Package> packageOptional = distCache.getPackage(loadRequest);
-
         return packageOptional.map(pkg -> {
             pkg.resolveDependencies();
             globalPackageCache.put(pkg);
