+++ compiler/ballerina-lang/src/main/java/io/ballerina/projects/internal/environment/LangLibResolver.java
@@ -15,13 +15,11 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-
-package io.ballerina.projects.env;
+package io.ballerina.projects.internal.environment;
 
 import io.ballerina.projects.Module;
 import io.ballerina.projects.Package;
 import io.ballerina.projects.PackageId;
-import io.ballerina.projects.Project;
 import io.ballerina.projects.SemanticVersion;
 import io.ballerina.projects.environment.GlobalPackageCache;
 import io.ballerina.projects.environment.ModuleLoadRequest;
@@ -36,36 +34,25 @@ import java.util.List;
 import java.util.Optional;
 
 /**
- * Default Package resolver for Ballerina project.
+ * Resolves lang lib packages.
  *
  * @since 2.0.0
  */
-public class DefaultPackageResolver extends PackageResolver {
-    private final Project project;
+public class LangLibResolver extends PackageResolver {
     private final Repository distCache;
     private final GlobalPackageCache globalPackageCache;
 
-    public DefaultPackageResolver(Project project, Repository distCache, GlobalPackageCache globalPackageCache) {
-        this.project = project;
+    public LangLibResolver(Repository distCache, GlobalPackageCache globalPackageCache) {
         this.distCache = distCache;
         this.globalPackageCache = globalPackageCache;
     }
 
+    @Override
     public Collection<ModuleLoadResponse> loadPackages(Collection<ModuleLoadRequest> moduleLoadRequests) {
         // TODO This is a dummy implementation that checks only the current package in the project.
         List<ModuleLoadResponse> modLoadResponses = new ArrayList<>();
-        Package currentPkg = this.project.currentPackage();
         for (ModuleLoadRequest modLoadRequest : moduleLoadRequests) {
-            Module module = null;
-            if (currentPkg.packageName().equals(modLoadRequest.packageName())) {
-                module = currentPkg.module(modLoadRequest.moduleName());
-            }
-
-            if (module == null) {
-                // Try to get from the already loaded packages
-                module = loadFromCache(modLoadRequest);
-            }
-
+            Module module = loadFromCache(modLoadRequest);
             if (module == null) {
                 module = loadFromDistributionCache(modLoadRequest);
             }
@@ -79,10 +66,8 @@ public class DefaultPackageResolver extends PackageResolver {
         return modLoadResponses;
     }
 
+    @Override
     public Package getPackage(PackageId packageId) {
-        if (project.currentPackage().packageId() == packageId) {
-            return project.currentPackage();
-        }
         return globalPackageCache.get(packageId);
     }
 
@@ -98,10 +83,12 @@ public class DefaultPackageResolver extends PackageResolver {
                 return null;
             }
             SemanticVersion latest = findlatest(packageVersions);
-            loadRequest = new PackageLoadRequest(loadRequest.orgName().orElse(null), loadRequest.packageName(), latest);
+            loadRequest = new PackageLoadRequest(loadRequest.orgName()
+                    .orElse(null), loadRequest.packageName(), latest);
         }
 
         Optional<Package> packageOptional = distCache.getPackage(loadRequest);
+
         return packageOptional.map(pkg -> {
             pkg.resolveDependencies();
             globalPackageCache.put(pkg);
